+++ addOns/spiderAjax/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Changed
+- Help to show that the Client Spider is now the recommended option for modern apps.
 
 ## [23.31.0] - 2026-05-21
 ### Added